A sample of the Drip Marketing Engine
A simple drip marketing engine written in SuiteScript.  This script can
be deployed multiple times, with each deployment passing in the id for
a customer saved search and the id for a template to send.  The engine
will send an email with the contents of the template to each customer
within the saved search.